% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promille.R
\name{compute_alcohol_concentration}
\alias{compute_alcohol_concentration}
\title{compute alcohol concentration}
\usage{
compute_alcohol_concentration(alcohol_mass, distribution_factor, mass)
}
\arguments{
\item{alcohol_mass}{in g}

\item{distribution_factor}{whatson distribution factor}

\item{mass}{in kg}
}
\value{
alcohol concentration
}
\description{
compute alcohol concentration
}
