% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promille.R
\name{compute_alcohol_mass}
\alias{compute_alcohol_mass}
\title{compute alcohol mass}
\usage{
compute_alcohol_mass(volume, alcohol_percentage, alcohol_density = 0.8)
}
\arguments{
\item{volume}{volume (of alcohol) in ml}

\item{alcohol_percentage}{permille amount of alcohol}

\item{alcohol_density}{in g/cm3}
}
\value{
alcohol mass in grams
}
\description{
compute alcohol mass
}
