% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promille.R
\name{compute_distribution_factor}
\alias{compute_distribution_factor}
\title{compute distribution factor}
\usage{
compute_distribution_factor(total_body_water, mass, blood_density = 1.055)
}
\arguments{
\item{total_body_water}{total body water (unitless)}

\item{mass}{in kg}

\item{blood_density}{in g/cm3}
}
\value{
distribution factor
}
\description{
compute distribution factor
}
