% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promille.R
\name{compute_total_body_water}
\alias{compute_total_body_water}
\title{compute total body water}
\usage{
compute_total_body_water(sex, age, height, mass)
}
\arguments{
\item{sex}{person's sex male or female}

\item{age}{person's age in years}

\item{height}{person's height in cm}

\item{mass}{person's mass in kg}
}
\value{
total body water (unitless)
}
\description{
compute total body water
}
